pyproject.toml
[project]
name = "aitool-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.49.0",
    "pydantic>=2.12.4",
]
